requirements.txt
python-whois
flask
beautifulsoup4
regex
requests
ipwhois
python-dotenv